requirements.txt
africastalking==1.1.5
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2==2.7.6.1
requests==2.21.0
schema==0.6.8
SQLAlchemy==1.2.15
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
